+++ compiler/src/main/java/info/teksol/mc/mindcode/compiler/optimization/LoopUnroller.java
@@ -421,14 +421,14 @@ class LoopUnroller extends BaseOptimizer {
         AstContext newContext = loop.existingParent().createChild(loop.existingNode(), AstContextType.BODY);
         int insertionPoint = firstInstructionIndex(loop);
         for (LogicList list : initContexts) {
-            insertInstructions(insertionPoint, list.duplicateToContext(newContext));
+            insertInstructions(insertionPoint, list.duplicateToContext(newContext, false));
             insertionPoint += list.size();
         }
 
         // Put in copies of the body
         for (int i = 0; i < loops; i++) {
             for (LogicList list : iterationContexts) {
-                insertInstructions(insertionPoint, list.duplicateToContext(newContext));
+                insertInstructions(insertionPoint, list.duplicateToContext(newContext, false));
                 insertionPoint += list.size();
             }
         }
@@ -470,16 +470,16 @@ class LoopUnroller extends BaseOptimizer {
         for (int i = 0; i < leading.size(); i++) {
             LogicList leadingCtx = removeLeadingIteratorInstructions(contextInstructions(leading.get(i)));
 
-            insertInstructions(insertionPoint, leadingCtx.duplicateToContext(newContext));
+            insertInstructions(insertionPoint, leadingCtx.duplicateToContext(newContext, false));
             insertionPoint += leadingCtx.size();
 
-            insertInstructions(insertionPoint, body.duplicateToContext(newContext));
+            insertInstructions(insertionPoint, body.duplicateToContext(newContext, false));
             insertionPoint += body.size();
 
             LogicList trailingCtx = contextInstructions(trailing.get(i));
             if (!trailingCtx.isEmpty()) {
                 // Skips labels/multilabels, copies the rest
-                LogicList copy = trailingCtx.transformToContext(newContext,
+                LogicList copy = trailingCtx.transformToContext(newContext, false,
                         ix -> ix instanceof LabelInstruction || ix instanceof MultiLabelInstruction || ix instanceof ReturnInstruction
                                 ? null : ix);
                 insertInstructions(insertionPoint, copy);